requirements.txt
mcp>=1.2.0
pypdf>=5.0.0
google-genai
pandas>=2.0.0
openpyxl>=3.1.0
python-telegram-bot>=21.0
requests>=2.31.0
